fix(a11y): stepper-completed step does not have accessible notification after completion (#1334) (#1348)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

In the Stepper, when the step is complete, keyboard focus moves to the
next step but there's no indication that the previous step is completed
by the screen reader.
<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: [CDE-699](https://jira.eng.vmware.com/browse/CDE-699)

The screen reader will announce that the current step is complete (or
failed) and move focus to the next step or stay in the same step in case
of an error.

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

---------

Co-authored-by: Andrea Fernandes <andreaf1@vmware.com>

---------

Co-authored-by: Andrea Fernandes <andreaf1@vmware.com>