fix(modal): fall back to modal id for dialog `aria-labelledby` attribute (#1403)
If you set `[clrModalLabelledById]` to an empty string, the
`aria-labelledby` attribute of the modal dialog element is set to an
empty value.

This behavior can be seen in the modal stories since the input is
defaulted to an empty string. We can't default it to the modal id
because that is not known until the component is instantiated.

Issue Number: CDE-2050

If you set `[clrModalLabelledById]` to an empty string, the
`aria-labelledby` attribute of the modal dialog element is set to the
modal id.

 (cherry picked from commit 6493dc3a6f34261faf34bfd6c8ab98784ffe65e1)