feat(modal): [MVP] add side panel (#1330)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] Tests for the changes have been added (for bug fixes / features)
- [x] Docs have been added / updated (for bug fixes / features)
- [x] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [ ] Bugfix
- [x] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: CDE-1835

## What is the new behavior?

A side panel component, implementationally based on the modal, has been provided

## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information
This MVP is a wrapper around the modal.
It covers the basic dialog functionality that's on-par with the modal.
A future, extended version of the component, would also address, these
cases which are not covered by this MVP:
- inline sidebars
- alternative positions - down, left
- side by side to the content
- side by side stacking of multiple sidebar dialogs
- etc.

The current API does not obstruct the above future goals.

---------

Co-authored-by: GitHub <noreply@github.com>