fix(a11y): return focus to datagrid detail pane toggle button after pane closed in zoom more than 250% (#1401)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?
In browser zoom more than 250%, when we open the details pane and close
it , the keyboard focus is not returned back to the button that was used
to open it. This is due to the fact that when the page is zoomed in, the
button is not visible on the page, and is available only once the detail
pane is closed.
<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: [CDE-1205](https://jira.eng.vmware.com/browse/CDE-1205)

## What is the new behavior?
Wait for the detail pane to close and then manually focus on the button
used to open it.
## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information

Co-authored-by: Andrea Fernandes <andreaf1@vmware.com>
(cherry picked from commit a6f345076bf0c5ba2e5eed142a562ad44dd4462a)