fix(datagrid): use proxy for detail context change (#1446)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

Angular deprecated replacing the full context for `EmbeddedViewRef`
which is used for datagrid detail and they suggest using a Proxy which
is also used in `NgTemplateOutlet`. Introduced with
https://github.com/angular/angular/pull/51887 and reported internally
that it create problems in v17.
`Swapping out the context object for EmbeddedViewRef is no longer
supported. Support for this was introduced with v12.0.0, but this
pattern is rarely used. There is no replacement, but you can use simple
assignments in most cases, or Object.assign , or alternatively still
replace the full object by using a Proxy (see NgTemplateOutlet as an
example).`

- [ ] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [X] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: N/A

## What is the new behavior?

## Does this PR introduce a breaking change?

- [ ] Yes
- [X] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information