fix(a11y): datagrid detail pane has role dialog with no aria label (#1405)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] Tests for the changes have been added (for bug fixes / features)
- [x] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?
`aria-describedby` is used on the datagrid detail pane instead of
`aria-labelledby`
<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: [CDE-645](https://jira.eng.vmware.com/browse/CDE-645)

## What is the new behavior?

- By default if the user does not add anything: `aria-labelledby` will
be added by the library with the id of the modal title.
- The user can set additional `aria-labelledby` (using the
[clrDetailAriaLabelledBy] input) which could have a value referencing
other text on the page , it can have multiple space separated id's. In
this case the default title id will be added along with the user defined
IDs.
- The user can add `aria-label` (using [clrDetailAriaLabel] input), in
this case the aria-label provided will be used.
- If both `aria-label` & `aria-labelledby` is added by the user, then
only `aria-labelledby` will be added to the modal.

## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information

---------

Co-authored-by: Andrea Fernandes <andreaf1@vmware.com>