fix(a11y): vertical nav - remove incorrect usage of links at top level for the vertical nav story (#1455)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [x] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?
This is a bug in the vertical nav routing story, and not the vertical
nav component. Links are used in the top level for the nav with routing
story which results in an independent accordion icon button being added
to the DOM that has state but no label.
The top-level entries should not be links. That usage is incorrect and
breaks the component.

<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: [CDE-1390](https://jira.eng.vmware.com/browse/CDE-1390)

## What is the new behavior?

## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information
This will fix related bug :
[CDE-1391](https://jira.eng.vmware.com/browse/CDE-1391)

---------

Co-authored-by: Andrea Fernandes <andreaf1@vmware.com>
Co-authored-by: GitHub <noreply@github.com>