fix(spinner): override wrong tokens mapping (#1500)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [x] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [x] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?


![image](https://github.com/user-attachments/assets/f50088df-935f-4bad-a7a1-fba1c1d1d084)


<!-- Please describe the current behavior that you are modifying, or
link to a relevant issue. -->

Issue Number: CDE-1513

## What is the new behavior?


![image](https://github.com/user-attachments/assets/4bd6d173-852e-4160-ad4e-90ebd0d0315e)


## Does this PR introduce a breaking change?

- [ ] Yes
- [x] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information

Spinner inverse tokens were correct but their mapping to the global ones
is wrong. This overrides the wrong mapping and these tokens are not used
in other places so we are safe with the fix.

Update: after discussion with @valentin-mladenov and design team we've
found out that two more tokens have incorrect mapping, so we are adding
them too in this PR.

Another improvement that the PR includes is providing a more meaningful
demo for the inverse spinner (with proper background) in the demo app
and the storybook

---------

Co-authored-by: GitHub <noreply@github.com>