fix(wizard): improve screen reader text for stepnav items (#1558)
## What is the current behavior?

- For uncompleted steps, the wizard step nav reads "1 [step title]". I
doesn't read "step" before the step number.
- For completed steps, the wizard step nav reads "complete 1 [step
title] completed". The first "complete" is not translated, and the
second one is redundant.
- For steps with an error, the wizard step nav reads "error 1 [step
title] error". The first "error" is not translated, and the second one
is redundant.

Issue Number: CDE-2292

## What is the new behavior?

- For uncompleted steps, the wizard step nav reads "step 1 [step
title]".
- For completed steps, the wizard step nav reads "completed step 1 [step
title]".
- For steps with an error, the wizard step nav reads "error step 1 [step
title]".

## Does this PR introduce a breaking change?

The "Step" text will need to be translated. We don't count new common
strings as breaking changes.