fix(label): allow signposts inside labels (#937)
## PR Checklist

Please check if your PR fulfills the following requirements:

- [ ] Tests for the changes have been added (for bug fixes / features)
- [ ] Docs have been added / updated (for bug fixes / features)
- [ ] If applicable, have a visual design approval

## PR Type

What kind of change does this PR introduce?

<!-- Please check the one that applies to this PR using "x". -->

- [X] Bugfix
- [ ] Feature
- [ ] Code style update (formatting, local variables)
- [ ] Refactoring (no functional changes, no api changes)
- [ ] Build related changes
- [ ] CI related changes
- [ ] Documentation content changes
- [ ] Other... Please describe:

## What is the current behavior?

Currently if you put signposts inside labels they are not working due to
that `<label>` be default is spreading a click event to its children and
that's causing the signpost to automatically close and not appear.
**Example:**
If you have a label attached to an input or input inside a label and
click the label - the input will get focused.

Issue Number: CDE-9 and #271 

## What is the new behavior?

Signposts are now working inside `<label>`.

## Does this PR introduce a breaking change?

- [ ] Yes
- [X] No

<!-- If this PR contains a breaking change, please describe the impact
and migration path for existing applications below. -->

## Other information

---------

Co-authored-by: Kevin Buhmann <kbuhmann@vmware.com>