feat(wizard): add in-page wizard (#1709)
## What is the current behavior?

Wizards must be shown in a modal. We have support for "inline" wizards,
but those intended to be shown in an iframe that's in a modal. As such,
inline wizards do not fill the whole page, use F-Pattern buttons, or
prevent close.

## What is the new behavior?

A wizard can be use in "in page" mode which can fill the page, uses
F-Pattern buttons, and cannot be closed.

CDE-2578